+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+### Fixed
+
+- Fixed Whisperx diarization pipeline initialization
+- Fixed Whisperx language detection
+
 [1.9.0] (2025-06-29)
 --------------------